+++ apps/modern-component-data-fetch/host/modern.config.ts
@@ -1,11 +1,8 @@
 import { appTools, defineConfig } from '@modern-js/app-tools';
-import { moduleFederationPlugin } from '@module-federation/modern-js';
+import { moduleFederationPlugin } from '@module-federation/modern-js-v3';
 
 // https://modernjs.dev/en/configure/app/usage
 export default defineConfig({
-  runtime: {
-    router: true,
-  },
   server: {
     ssr: {
       mode: 'stream',
@@ -13,9 +10,7 @@ export default defineConfig({
     port: 5001,
   },
   plugins: [
-    appTools({
-      bundler: 'rspack',
-    }),
+    appTools(),
     moduleFederationPlugin({
       fetchServerQuery: { extraQuery: true },
     }),