+++ apps/shared-tree-shaking/with-server/host/CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Patch Changes
 
-- @module-federation/modern-js@0.24.1
+- @module-federation/modern-js-v3@0.24.1
 
 ## 0.1.35
 
 ### Patch Changes
 
 - Updated dependencies [f46cdd6]
-  - @module-federation/modern-js@0.24.0
+  - @module-federation/modern-js-v3@0.24.0